+++ src/adapters/ptm-exchange-adapter.ts
