+++ src/adapters/variation-graph-adapter.ts
