--- src/loadComponents.ts
@@ -1,10 +1,0 @@
-const loadComponent = function (
-  name: string,
-  elementConstructor: CustomElementConstructor
-) {
-  if (!customElements.get(name)) {
-    customElements.define(name, elementConstructor);
-  }
-};
-
-export { loadComponent };
