+++ src/utils.ts
@@ -10,3 +10,12 @@ export const renameProperties = (features) => {
     };
   });
 };
+
+export const loadComponent = function (
+  name: string,
+  elementConstructor: CustomElementConstructor
+) {
+  if (!customElements.get(name)) {
+    customElements.define(name, elementConstructor);
+  }
+};
